Fix event delegation with React 17+
Documentation: https://legacy.reactjs.org/blog/2020/08/10/react-v17-rc.html#fixing-potential-issues

The recommended fix is to use `{ capture: true }` when defining the events.

I went through the various `document.addEventListener` calls and updated the ones that I found were not working.

If the `options` parameter is a boolean, this is used as the `capture` option.

I also unified the various calls to handle browsers with passive events the same way as browsers not supporting it.