Add view spec for error_message i18n pluralization
The interaction of the Rails translation helpers and the i18n gem can
lead to a scenario where on certain version combinations, a locale which
has the `other` pluralization key defined for the
`generic.validation_errors` translation, but did not have the `one` key
defined would lead to a pluralization error when rendering this partial.

This spec should catch regressions in this handling of this case so that
as we update rails and i18n gems we are more certain that we aren't
breaking this behavior.