Properly polyfill Intl
The `intl` package has been deprecated, so let's replace it with FormatJS polyfills.

It also allows to properly load the locale data for every locale, before it always loaded `en` from the polyfill.
This created a Webpack chunk for every locale, but this is needed if we want per-locale loading.

I chose to only enable the polyfill for `Intl.PluralRules`, as the browser support for the others is good noaways. `RelativeTimeFormat` will need to use its own polyfill once we use it in the codebase, but as its not used for now, its commented.
I left the other polyfills commented out, if we notice they are really needed somehow.

The change in `tsconfig.json` is needed for dynamic imports, it has no impact for Webpack/Babel