Upgrade to react-router v5.
The biggest (and only) change is `this.context.router` not existing anymore (it switched to the new React hook context).

It has been either removed when not used, or replaced by `withRouter` + props

In one case, `withRouter` messed up with `ref`, so I used a wrapper function component to be able to call `useLocation()`

Some components are used outside of a router context, and `withRouter` will fail if called outside of a context, so I introduced a `withOptionalRouter` that will not set the router props if outside of a context, but those props existence needs to be checked in the component using it.