Improve follow recommendations SQL query
A server admin reported that this query took 29 hours to run on their instance,
using PostgreSQL 14 with up-to-date vacuum & statistics.

After some investigation, we found out that the plan was very wrong and
resulted in 300B+ rows being generated in a temporary table.

Switching from a `JOIN` to `NOT EXISTS` fixes the issue and the query is
now behaving as expected.

This has been tested on PostgreSQL 12 on a large server, and performs the
same as with the `JOIN` on this server.

Fixes #25191