Speed-up on `Rack::Attack` spec
Previously we were looping through `limit` number of requests to
increment the counter 1-by-1 and get the `Rack::Attack` data store (the
Rails cache in test mode) to a place where it would see the next request
as needing to be rate limited, and we could then assert that this
happens correctly.

This is slow because in the background all the associated http requests
and processing for every action needs to actually happen `limit` number
of times, all just to get the counter incremented.

This changes the spec so that on the `limit` times loop instead of doing
a full http request/response cycle, we just directly increment the
cache, and then assert that the next request after that is either not
blocked (in the lower case) or is blocked (in the higher case).

Before: 216 action requests, 569 DB queries, ~3.9 seconds
After: 11 action requests, 79 queries, ~0.7 seconds