Load `rspec-rails` gem in `test` + `development`
The rails mailer previews only run in development mode, and they default
to looking in `test/mailers/previews` as a load path for previews.

The `rspec-rails` gem sets the path to `spec/mailers/previews`, but was
only being loaded in the test environment.

This change makes it so that `rspec-rails` is loaded in both test and
development environments, and thus sets the preview load path correctly
in the development environment.