Ignore the brakeman `PermitAttributes` check
Avoids brakeman mass assignment warning.

This check includes a hard-coded reference to `account_id`, which we use
(safely) in a bunch of places.

In the event we introduced an unsafe permit, it would likely be caught
by other Mass Assignment checks, and we are only going to get false
positives from this particular check.