Simplify instance presenter view access
Previously, there were many locations which render the shared/og partial
and set up an @instance_presenter variable exclusively for use there.
There is also a memoizing helper method in mascot helper which does the
same thing.

This change converts the view to use the memoized result of that helper,
and removes all the i-var set up from the controllers.