Fix the crossorigin attribute
We are seeing some warnings about `crossorigin="anonymous"` being inconsistent between the preload requests and the real load.
This makes the preload useless as the browser ignores it.

With this fix, I configure Webpack to add the attribute when it generates `<script>` tags (dynamic imports).

But Webpack only inserts it if the source is loaded from an external domain (otherwise the `crossorigin` attribute does nothing), so we need to only add it in the preload tag when an external domain is configured for the assets.