Coverage for OmniauthCallbacks controller
Adds coverage for the "happy path" of finding a user with matching email
and uid via the openid_connection omniauth provider.

This spec only runs when the OIDC_ENABLED env var is set to `true` and
the OIDC_SCOPES value is present. These are required at minimum to
load/configure the provider.

This adds an omniauth test mock, which for now is hard coded to values
usable by this one specific example.