Restore console behavior in `test` env
A previous commit added the profiling gems to both test/development
environments: https://github.com/mastodon/mastodon/pull/25383

During this change, we added some rspec type auto inference code, and as
part of that change we moved the fabricator spec to sit in the
spec/fabricators/ directory.

This had an unintended side effect of breaking the use of `Fabricate`
from with in the Rails console.

When rails/bootsnap loads all the fabricators, that file requires
`rails_helper` which throws an error.

This change moves the fabricator spec out of the fabricator-defining
dir, and into another directory.

This preserves the behavior of this file as a spec, and fixes the broken
console behavior.