Change S3 checksum mode to be disabled by default
This removes the `S3_DISABLE_CHECKSUM_MODE` env variable and introduces the
`S3_ENABLE_CHECKSUM_MODE` env variable instead.

The reason is that it seems that more people use S3-compatible providers which
do not implement this feature (and thus error out if enabled) than
S3-compatible providers which do implement it (which might be AWS S3 only).