Consider shown and pending status in explore prompt calculation
The explore prompt had been picking between shown and pending statuses
when deciding whether to display the explore prompt, which prompts you
to follow more people if your timeline is not very busy.

Considering just the pending statuses would mean that sometimes, if you
had posted something which was now a pending status, that would be
filtered out (as the explore prompt calculation filters out our own
status) leaving no statuses, and thus show the explore prompt, even if
there is plenty of activity in the shown statuses.

Instead, we should be considering the first 20 of both the pending and
shown statuses for this computation; we need to concatenate the lists.

Fixes #27197, fixes #27362