Simplify column headers and use automatic back buttons
I do not have the context to understand why it is done this way (it dates back to 2017), but I am not sure we still need two `<ColumnHeader>` components.

We can simplify the code this way, and the columns that were using `<ColumnBackButtonSlim>` have their title properly displayed now, it was hidden before.