Only check that we are logged in when really needed
I made a mistake in #26559 and inadvertently added a check when the reducer runs, which happens on page load.

This needs to only be checked when those specific actions are dispatched, which only happen when you are logged in.

Added a system check to ensure the UI loads when not logged in.

Fixes #27697