Disable Babel polyfills in dev
In #27333 I added useBuiltIns: "usage", which causes Babel to inject polyfills into those files. The polyfills are injected using ESM import symtax, which disables the handling of export. (CommonJS exports).

This causes the build in dev to fail, because in dev Babel does not convert the exports to CJS (not sure why).

I tried various things but I am unsure of why this happens. A simple fix is to no longer have Babel inject polyfills in dev, which was the previous behaviour.