Remove explore prompt
In #25267 a prompt was added to encourage people to follow more accounts
and hashtags if there weren't enough posts or variety when viewing your
home timeline.

This prompt has been showing up at a lot of inappropriate times on busy
home timelines, leading to a series of bugs and PRs about it showing up
when it shouldn't. See for example #25569 #25570 #25579 #27211 #27197

In #27726 and associated PR #27728 I discuss how I don't think that
there really can be a reasonable fix for this prompt showing up on
timelines which contain posts that have been delayed for whatever
reason (outgoing or incoming queue on servers being behind, for
example), without substantially increasing the number of posts we load
initially or implementing the feature on the server side. Additionally,
the dismissal state being stored purely client side means that the
prompt can frustratingly keep showing up even if you dismiss it if you
move between browsers, computers, private tabs, etc.

So it seems that to truly implement this feature well, it can't be a
purely client side feature; it needs some amount of API and server side
implementation. That brings up the question of whether this feature is
worth that complexity. This feature is prompting people to follow more
people or hashtags because their home timeline is a bit slow; but maybe
some people like it that way, or follow mostly the local or global
timelines. One of the reasons that some people prefer Mastodon is that
it doesn't aggressively push discovery features, while this is a fairly
aggressive discovery feature, and without a bunch of extra effort can
come up in quite inappropriate contexts.

So this PR proposes just removing the feature. Fixing the feature,
probably by moving more of the implementation server side, and possibly
making it a bit less aggressive, might be another option, but I wanted
to propose just removing it to get a feel for people's thoughts.