OAuth: Enable refresh tokens
This introduces the ability for an application to request a new access token on demand, allowing for access token rotation. This does not introduce expiry of access tokens, but does allow for future introduction of access token expiry, as developers would be able to upgrade to supporting usage of refresh_tokens over the next release or few.