Use composite primary key for `PreviewCardsStatus`
This was previously set to nil because the class has a composite primary
key and Rails did not support that DB design. Rails now does have
(limited) supported for composite primary keys, which is sufficient for
our one use case: letting the destroy callbacks on Status and
PreviewCard correctly connect to their associated records and run.

This change also refactors what was previously a `before_destroy`
callback to just use the built-in `dependent: :destroy` approach on the
association.

The general SQL is slightly different, primarily because the previous
method was doing a `delete_all` even though this is a has one. The new
query continues to delete the connect PreviewCardsStatus record for the
Status being destroyed, but issues a query with the full composite key
of the connected record now.