Handle CLI failure exit status at the script level
Previously the commands would output error messages and `exit(1)`
directly from the running code when there was an error condition or
invalid option.

This was a frequent source of false negative spec run issues where the
rspec `output ... to_stdout` block would wrap an `exit` call in a way
where the overall suite would keep running as though no errors had
occurred, but any specs for the class which threw the exit would not
actually have run. This both gives a false sense of spec status, and
also makes it hard to reason about and improve the existing code and
specs.

Separately, this is a nice DRY up of the classes, as what was some
previously multi-line condition checks are now just one-line guard
clause type checks.

The main `tootctl` script has been updated to properly handle the error
catching, and continues to output a message (colored red) with the error
code as before -- so from the perspective of a CLI user, nothing has
changed.