Solve `Abc/*` metrics for `db/*migrate*` files
These files were the only two necessary exclusions for these `Abc/*`
rules, and they are also exact copies of each other, minus the class
name.

The change here preserves the original logic and order of the migration,
and just moves the finders and logic loops to private methods, all of
which are called from the main `up` method in the migration.