Remove `add_column_with_default` migration helper
In a previous commit we removed the bulk of this method:
https://github.com/mastodon/mastodon/commit/4aa05d45fc45b797e0a3b7d40885ef89c2161a73#diff-006ad05199b268fb6057c942a06c939ad2f1da6f43cdc1260e5cfb1806560cd5L423-L458

With the postgres 12+ requirement, we didn't need the conditional
behavior anymore -- however, with all that removed, the method is just a
thin wrapper (with almost the same signature) as the built-in framework
method to do the same thing.

This updates all the migrations which were using that helper method to
just use the method directly, and modifies their signatures to pass in
`null` instead of `allow_null` (which the method was passing along to
underlying as a `null` option).

I believe that all generated SQL and migration logic should not change
here -- just a further clean-up of this method after the initial work.