Add a version comment to the emoji preval-ed file so it is regenerated
This is a follow up on #28404

In the PR above, the emoji data has been changed. But `emojii_compressed.js` is using `babel-preval`, so it is executed when processed with Babel, and then Babel will cache the resulting code.

This causes the frontend to no longer load after deploying the change, because Babel did not detect any change in `emoji_compressed.js` (the real file), so it re-used it from its cache, but the actuel content that needs to be consumed by Webpack and appear in the final bundle is different (it contains the new emoji data).

Adding this comment will change the checksum for `emoji_compressed.js`, which should invalidate Babel's cache and force the file to be executed again on the next build.