Change icon SVG imports to use a default import
This change is here to prepare for the Vite migration (and it makes the import lines cleaner).

This new syntax (default import, `?react` at the end) is the one that will be used by Vite. Replacement has been done using a regexp search/replace.

To get a working default import for SVG components, Webpack needs to not have another loader process the SVG files (because this other loader will replace the default export). I tried to configure the loaders so `.svg?react` files are handled by `svgr` and are ignored by the file loader, but I did not manage to make it work (Webpack 4 does support `resourceQuery`, but not with a negative matching). So I went with an `exclude` on the `file-loader` rule, matching the `include` in the svgr rule.

This configuration basically ignores the `?react` suffix, but it will be there so those imports will continue to work with Vite once we migrate.