Streaming: Rework websocket server authentication & initialisation code
This moves us away from the deprecated verifyClient mechanism, towards using a
custom 'upgrade' event handler as recommended by the author of the ws module.

In the future, we could even move towards using a custom WebSocket subclass to
add in additional methods and data directly to the WebSocket instance.