Use a helper to avoid using `media/` asset prefixes
This is in preparation for the Vite migration.

Once migrated, there is no longer a concept of a "pack". I chose to replace this with "frontend asset", which is more explicit.

With Vite, we will also no longer have the `media/` prefix, so it makes sense to have it added in the helper, as the change for Vite will be in a single place