Remove `eslint-plugin-prettier`
It's usage is no longer recommended as ESlint should not be used anymore for any formatting rules.

More details in this article: https://www.joshuakgoldberg.com/blog/you-probably-dont-need-eslint-config-prettier-or-eslint-plugin-prettier/

This change should speed up ESLint, and formatting issues will no longer show as ESLint warnings when editing, which often happens when developping. It is recommended to configure your editor to format using Prettier on auto-save, otherwise files will be fixed automatically before commit, or when `yarn fix:js` is ran.

The main change here is in the lint/fix scripts, as Prettier needs to be called separately.