Fixed the toggle emoji dropdown not working
The toggle function gets broken by the `handleDocumentClick` in EmojiPickerMenuImpl.

When you click the emoji picker button, it makes the dropdown active, but then, whenever you click outside of the emoji dropdown (i.e. the emoji picker button, in this case), it triggers the hide function from outside of the toggle system, which causes the toggle function to set active to true again.