Use full snowflake range in `admin/metrics` classes
The snowflake ID works by mapping a timestamp to a numeric value to use
for database IDs, with or without randomness.

When we are gathering IDs to use as the beginning or end of a report
range, we want to include all statuses after the smallest possible ID
and before the largest possible ID.

The previous code could lead to "partial day" status ID ranges, since
the passing in a value which was not already nudged to the beginning/end
of a day would only include some of the possible IDs for that day.

The change here forces the dates to beginning/end of day before asking
snowflake for the IDs.