Set rubocop `Style/HashSyntax` option `EnforcedShorthandSyntax: either`
Starting with ruby 3.1, the enforced rubocop default for hashes is to
prefer the shorthand syntax which omits hash values when the local var
names match the hash keys.

For example, in this scenario:

    Fabricate(:status, account: account, poll: poll)

The cop is going to enforce this style:

    Fabricate(:status, account:, poll:)

This works the same way, with the values in variables matching the key
names passed in.

We have thousands of violations of this cop, so lets set this value to
allow `either` style.