Makes the star icon rotate around its actual centre axis
The 55% value is slightly too low visually, meaning the star wiggles a bit. Measured 52% (rounded from 52.0833333) with the icon file opened in a vector editing software from the top of the icon canvas to the actual geometrical centre of the star shape. Tried it out in dev tools, and it’s smoother :)