Remove the access token from Redux and identity context
This is not mutable once initially loaded, and it does not make sense to store it into Redux, and this makes it easy to expose it (when dumping the Redux content for troubleshooting, dev extension…).

It is also very unpractical to require access to `store.getState` to make an API request where the only thing you need is the token.

This PR:
- disallows `access_token` to be stored in Redux
- removes `access_token` from the identity context (this was not used)
- creates a new `getAccessToken` accessor from the initial state
- changes all the places using `accessToken` to use the new function
- changes the `api(getState)` interface to `api()`, and update all the callsites