Retain unconfirmed users longer (1 week)
This is especially useful on instances with a CAPTCHA configured, otherwise if people that email you because they can not complete the captcha you might be too slow to process the request and the account might have already be deleted. 7 days instead of 2 provides some leeway.

I also extracted the durations to constantes to make them more explicit.