Remove unused `Extractor#extract_cashtags_with_indices` method
Method originally added to override the method in the underlying
twitter-text gem, as a way to avoid extracting cashtags from text:
https://github.com/mastodon/mastodon/pull/3275/files

Subsequently, the top-level method which calls it was also overridden:
https://github.com/mastodon/mastodon/pull/17828/files#diff-bbea6c6a6594c95c3de43213d551aff9b9b329001210a584d39dca9966ea0e1aR8-R21

Since that second change, we no longer need the original override.

Left the spec in place so that any future changes to that calling method
don't break the behavior.