Change PostCSS config to allow logical properties
We use the `insert-*` properties in the codebase, but `postcss-preset-env` replaces them with their LTR equivalent for compatibility. This is well supported in recent browsers, so let's disable the plugin so they are used in the final CSS.

Autoprefixer is now automatically called by `postcss-preset-env`, so we can remove it from our config.