Add a border above the timeline hint for statuses with replies
I tried to do it in pure CSS, but this does not work as the sibling divs do not have a class (they are `<HotKeys>` wrappers).

The border is only added when needed. I checked the other places the hint is displayed, and there is already a border here (due to the previous component having it).

Fixes #31385