Fix log out from user menu not working on Safari
The main issue here is that the form was not submitting. Adding a delay after adding it to the DOM seems to fix it, but this is brittle.

As we want to move away from form submits in the app to tighten the CSP policy, I chose to rewrite this to a direct request.

After a log out, the backend responds with a 302 request, redirecting to another page. This needs to be reimplemented in JS, but there is no way to get the `Location` of a redirect using `fetch()`. So I went to change the backend to respond with a JSON object to JSON log out queries, exposing the URL to return the user to.

Fixes #31383