Disable Chrome's search engine selection screen when running system specs
I refactored the options to be able to set some base one for both drivers. I think it makes sense to set the window size all the time, as our specs might rely on it.

In a Selenium issue (https://github.com/SeleniumHQ/selenium/issues/14343) they said that they do not want to change the default browser behaviour and each app should do it, so here we go.

This only applies when Chrome detects that you are in Europe I think, thats why it did not happen in CI.