Improve the list selection UI for notification requests
This should solve the usability problems on mobile, show less options to the user, while still allowing every operation.

I chose to rename the translation keys for consistency, as well as adding `…` to the menu items to indicate that there is a confirmation dialog, and clicking on it will not immediately do the action.