wip: Ensure datetime are emitted in RFC3339 format
The `datetime` type should be emitted in RFC3339 format.

Ensure that happens by adding an `ActiveSupport::TimeWithZone`
implementation with a `as_json` method that returns the value using
`rfc3339(3)`.

Replace existing uses of `.iso8601` with `.rfc3339` -- the format
string used is the same, so this is a no-op for external behaviour,
it makes the code match the documentation.

Add an example test. Since the existing tests don't hardcode the time
it was simplest to make sure the serialized value can (which should be
in RFC3339 format) can roundtrip through an RFC3339 parser.