don't start OTel traces with Redis spans
The Redis instrumentation picks up a lot of activity that does not have a parent context. (e.g. Sidekiq polling for jobs. App library methods that are not seen by auto-instrumentation using Redis as a cache.)

This change flips an option on the instrumentation to never generate a Redis span unless there is already another current span started by something else. This will reduce the volume of essentially-anonymous low-value Redis calls.