Use a system setting for the Referer policy
Follow up on #33214

Having this as a setting will allow exposing it as an admin setting.

There is a slight drawback here as reading the setting for every request makes a Redis call, but I am not sure if we have a way to cache it in memory for the Rails process for a few seconds. Not sure if this additional Redis query will have any impact at all.