introduce drag & drop handlers to search box
without modification, default browser behavior for dropping new text on top of pre-existing text in an input field is to insert the new text in between whatever characters the mouse is pointing to at time of drop.  Instead, by adding handlers, the pre-existing content can be overwritten/replaced.  The primary use-case is for accessing quoted toots, or referenced hashtags or accounts, in the feed by searching for the linked URLs & showing the original/expanded version of them in the search results (as opposed to clicking on the links themselves and having them open in a new tab, on a potentially different mastodon server)
Since the search box is designed to actively process its input for context-clues, after the text has been dropped, an Input event is triggered to bubble up from the search-box's new handleDrop method, that will cause the handleChange method to be invoked.  This strategy was adapted from https://stackoverflow.com/a/47409362