refactor handleChange to extract context from URLs
previously, the search box would only add new quickActions if raw `@` or `#` characters were found in the search string; but in the case of a drag and drop, or a copy-paste, of a URL into the search bar, no elements of the URL's path would be recognized for potential quickActions.
This change enables identification of possible hashtags or user accounts in a (presumed) mastodon URL; augmenting id by special input characters.