Update confirm_captcha check to use either method
Updated confirm_captcha check to use either method
available depending on whether it is mCaptcha or hCaptcha
enabled.

Currently doesn't work since we need to change it
to be compatible with an if/elsif statement. Adding it now
so that we don't forget update it during cleanup/changes
related to feedback from testing.