Swap order of translation restoration and service credit
Swapping them moves the "show original" link to the same location where the "translate" link was, increasing the likelihood that a user need not move the pointer to restore the original— but not guaranteeing it.

Further reasoning: As a language learner, it's educational to follow accounts posting in the language I'm learning. I use the "translate" feature to confirm my understanding or teach me vocabulary I don't know. I will sometimes swap between the original and the translated version multiple times. This small UX tweak dramatically decreases the chance that one must move their pointer at all, whereas before, the "show original" link was _always_ on the other side of the box and required pointer movement.