Reduce number of labels
For this minimal middleware we do not need the full number of
labels. And for queue time, that happens before a request hits
the app, the actual endpoints within the app are not relevant.

People wanting the full number of labels can always set
`MASTODON_PROMETHEUS_EXPORTER_WEB_DETAILED_METRICS`.