Update to linzer 0.7
In this release, linzer introduces a new API and fixes some
bugs. This commit updates the gem and the code to use the new
API.

Note that this fixes a bug that is still present in our FASP
implementation, so for a short while this will be incompatible
with them.