Integrate FASP follow recommendations
When a user requests follow recommendations, configured matching
FASP are queried in the background. The new "async refresh"
mechanism is used to inform clients that can then reload results
after a while.

This is not perfect in the sense that there is no guarantee that
recommendations from FASP will actually also be recommended by
the local server once they are known there. But it should still
improve results in most cases.