Remove "irrevocable" and "perpetual" from license
Per #35086, requiring users to grant an "irrevocable" and "perpetual"
license to content is a considerable overreach, and most similar
services have clauses that sever such grants once content is removed
from services.

Update the ToS template to reflect this, while adding clauses to allow
for incidental continued storage as part of caches or backups which may
not be stored for an indefinite period. Given the variety of ways in
which people may configure caching or backup, I'm not sure it would be a
good idea to commit to a time period in which caches and backups would
be deleted in a default ToS, so leave that period indefinite.

Additionally, clarify that due to the nature of federation, copies may
have been propagated to servers outside of the Administrator's control,
and such copies are not subject to the terms of this agreement. Since
this is simply an agreement between the user and the administrator of a
particular instance, I don't think that anything can be said about
agreements with other instances, but I think that it is worth mentioning
for clarity.

Note that I am not a lawyer; this is just an attempt to limit the scope
of the license grant somwhat, while still keeping the liability of the
administrator relatively constrained as long as they administer the
software in a fairly standard way and don't do anything to interfere
with the deletion of content. I am open to suggestions on alternate ways
of wording this.