Fix synchronous recursive fetching of deeply-nested quoted posts
Fixes #35572

Like other resources, quote posts can be recursively fetched, and the depth
is limited through the `request_id` mechanism.

However, unlike other resources, they are fetched synchronously, which is required
for a good user experience, but drastically changes the nature of the recursion,
and the global asynchronous recursion limit we use is far too high for synchronous
recursion.

This commit introduces a synchronous recursion depth limit to only fetch a maximum
of 2 nested quotes synchronously, leaving the rest to later jobs.