Experimentation with switching to Tanstack Router
This is not really working except for 3 screens (Home, Notifications, Explore).

Multi-columns management and all the login from `<UI>` has not been implemented yet.

Many components are still using `<Link>` or `withRouter` and will fail if renderer inside the new router

Some links have been converted to plain links and will need to be switched back

There are issues with loading content, because this is tied to the previous scroll restoration and is not triggering properly. All the data loading should probably be reworked completely to fit with the new router