Add flag to preserve cached media on cleanup
Adding the flag --keep-interacted to the "media remove" tootctl command
to preserve media cached from other servers if the related status has
interactions from local accounts.

For example, you may wish to delete all cached media from other servers
older than 3 months to save disk space on the local server. However, if
the status was bookmarked by a user they likely want to keep a copy
locally, preserving the media in case the remote server is closed or has
decided themselves to clear out all old user media.

This flag was proposed in issue #30449.

The change does not impact the local content retention setting, which
manages automated cleanup. A new setting would need to be added there in
a future improvement.

This means to make use of this change a server admin would have to set a
long content retention setting beyond which all remote media would be
cleared, and then use a shorter range with the tootctl command where
interacted media would be preserved with this new flag.

The flag is only to preserve media attached to statuses, and makes no
difference to the logic removing cached avatars and headers.